Audio memos: post-implementation hardening (reflect pass)
Four fixes out of the wave-3 reflection:

- Age-gate the persisted native action (15 minutes). Re-firing a
  crash-orphaned record request seconds later is the handshake's contract,
  but turning the microphone on days after the tap that asked for it is a
  surprise no user reads as their own action — a stale request is dropped,
  not delivered.
- End stale Live Activities on plugin load. A crash mid-recording left its
  activity counting on the lock screen forever (the orphan scan saves the
  audio, but nothing ended the activity). Nothing can be legitimately live
  at load, so end them all.
- Rename the Live Activity stop intent to StopRecordingLiveActivityIntent
  and mark it non-discoverable — the Shortcuts app was about to list two
  identical "Stop recording" actions; StopRecordingSiriIntent is the
  discoverable one.
- Extract use-staged-recording-ingest (the orphan scan) and
  use-native-record-action (the reconcile-then-handshake sequence, whose
  ordering rationale now lives on the hook) from the mobile provider,
  which shrinks to composition — the repo's established use-* hook shape.
  No behavior change; the provider suite passes unchanged.
